update calMonthlyInterest methods to return double; adjust calculations to use BigDecimal

+++ src/main/java/domain/amount/FixedDepositAmount.java
@@ -30,6 +30,6 @@ public class FixedDepositAmount implements LumpSumInvestmentAmount {
 
 	@Override
 	public double calMonthlyInterest(InterestRate interestRate) {
-		return interestRate.calMonthlyInterest(amount);
+		return interestRate.calMonthlyInterest(amount).doubleValue();
 	}
 }
