+++ src/main/java/domain/amount/MonthlyInstallmentInvestmentAmount.java
@@ -25,6 +25,6 @@ public class MonthlyInstallmentInvestmentAmount implements InstallmentInvestment
 
 	@Override
 	public double calMonthlyInterest(InterestRate interestRate) {
-		return interestRate.calMonthlyInterest(amount);
+		return interestRate.calMonthlyInterest(amount).doubleValue();
 	}
 }
